Websocket and Control Flow
Began adding websocket interactions. During this implementation,
discovered issues with flow. Fixed them for the most part, but still not
what was expected.

Forms still appear one after another, however it is done in the proper
manner this time, and the user can click cancel to back out to the
previous form.

As for the websocket code. Not much was actually added. Will start with
signin messages next update.