Websocket and XML implementation
Began inserting websocket functionality by introducing xml handling
code. This mostly includes a means of building xml strings for reading
and then parsing their parts into a dictionary to be used later.

Also suggested is an attribute called "error" to determine for instance
if a login fails on the server side. It could be paired with values such
as "username already exists" or "username does not exist" depending on
whether the user is trying to sign up or login. Other errors could occur
as well (passwords) and we should decide what the values for those
should be.