Access Message from Server Handled
Filled out code for what to do when an "access" message is received from
the server. Using the word access in descriptions to refer to the
combined login and signup actions.

Suggested xml attributes "dispName" and "ip" to refer to the values
required from the server. Updated appropriate documents in the
XML_EX_FILES folder.

Error messages received under the "error" attribute can be completely up
to the server side because the client side just echos it without any
need to interpret.