Create Conversation Form
Made a form and most of the flow control for creating a group
conversation without editing a current conversation.

Added an option in the conversation menu for added another person to the
current conversation.

Added textboxes to the add and remove contacts options so that
individual forms will not have to be made for them.