Client: Readability Changes and Generalized Error Calls
Moved the sending conversation creation messages to its own method to
improve readability of the delegate implementation.

I also forgot to mention that I added the ability to send generalized
error messages to the client anytime now. So long as the ' error="" '
attribute is included in an element and it doesn't use a predefined
element name (i.e. crConv, msg, login, content, etc).

So now there is specific errors:
<login error="username or password is incorrect">

and general errors:
<someName error="Ya done fudged up, kid.">

where once again "someName" is not one of the predifined keywords such
as login, msg, or crConv.